Adding new types of bullets and monsters.
Added new type of bullet for the second level, and new type of Monster
(called MonsterStrong) for the third level. Also improved the logic for
killing different types of Monsters. Basically when dealing with
MonsterMedium (the witch) you need to shoot her twice in order today, as
for MonsterStrong (the fireball) you need to shoot it three times. You
get 2 and 5 points for them respectively.